==> 4_Cell_Type_annotation.Rmd <==
---
title: "Cell type annotation"
output:
  html_document:
    df_print: paged
---


Libraries
```{r Librairies}
library(tidyverse)
library(dplyr)
library(clustree)
library(patchwork)
library(Seurat, lib.loc = "/opt/rstudio-server_conda/conda/envs/rstudio-server_4.1.0/lib/R/library/")
```

Import objects
```{r Objects}
SCT_obj <- readRDS("Objects/SCT_obj.rds")
Biomart_annotated <- read.csv("Analysis_tools/Biomart_annotated.csv")
```

Clear unused resolutions from metadata
```{r}
metadata_to_clear <- paste0("int_20_snn_res.", seq(0.2,2,0.2))
SCT_obj@meta.data[,metadata_to_clear] <- NULL
```

We used selected genes from the literature to characterize the various clusters
```{r Marker expression}
Genes_to_show <- c("Slc1a3", "Aqp4", "Gfap", "Vcam1", "S100a6", "Egfr", "Ascl1", "Top2a", "Mki67", "Dcx", "Robo2", "Slc17a6", "Pdgfra", "Il33", "Mog", "S100b", "Dnah11", "Acta2", "Flt1", "Pecam1", "Pdgfrb", "Des", "Vtn", "Fyn", "P2ry12", "Cd3e", "Clec4e", "Cox8b")

clusters_order <- 
  c("29", "4", "13", "17","20", "10", "5","12", "15", "8",
    "19","1", "3", "7","9", "30","27","16","24","23", "6","14", 
    "21", "28","0", "2","11", "18", "25", "31", "22","26","32")

SCT_obj$clusters <- factor(SCT_obj$clusters, levels = clusters_order)

Idents(SCT_obj) <- "clusters"
```


Based on this gene expression patterns, we assigned cell types manually for every cluster
```{r Cell type assignation}
cell_annotations <- c(
  "Astrocytes",
  "Astrocytes",
  "Astrocytes",
  "NSCs",
  "TAPs",
  "Cycling Prog.",
  "Cycling Prog.",
  "Cycling Prog.",
  "Cycling Prog.",
  "Cycling Prog.",
  "Neuroblasts",
  "Neuroblasts",
  "Neuroblasts",
  "Neuroblasts",
  "Neuroblasts",
  "Neurons",
  "OPCs",
  "MFOLs",
  "MFOLs",
  "Ependymal",
  "Endothelial",
  "Pericytes",
  "Pericytes",
  "Microglia",
  "Microglia",
  "Microglia",
  "Microglia",
  "Microglia",
  "Microglia",
  "Immune",
  "Immune",
  "Immune",
  "Immune"
)

names(cell_annotations) <- as.factor(clusters_order)

Idents(SCT_obj) <- "clusters"

SCT_obj <- RenameIdents(SCT_obj, cell_annotations)

SCT_obj <- AddMetaData(SCT_obj, 
                       metadata =  Idents(SCT_obj), 
                       col.name = "Cell_Type")
SCT_obj$Cell_Type <- factor(SCT_obj$Cell_Type, levels = unique(cell_annotations))

UMAPPlot(SCT_obj, group.by = "Cell_Type", label = T) + NoLegend()
```

Selecting NPC cells only
Cluster 29 was excluded from NPC-restricted analyses as it shared features from astrocytes and microglia
```{r NPC metadata addition}
#Creating NPC metadata for easy selection (using subset(object, NPC == "NPC"))
cell_annotations[c(1,16:33)] <- "Micro-environment"
cell_annotations[2:15] <- "NPC"

Idents(SCT_obj) <- "clusters"

SCT_obj <- RenameIdents(SCT_obj, cell_annotations)

SCT_obj <- AddMetaData(SCT_obj, 
                       metadata =  Idents(SCT_obj), 
                       col.name = "NPC")

#Creating NPC-restricted ordered cluster metadata
cell_annotations[c(1,16:33)] <- NA
cell_annotations[2:15] <- clusters_order[2:15]

Idents(SCT_obj) <- "clusters"

SCT_obj <- RenameIdents(SCT_obj, cell_annotations)

SCT_obj <- AddMetaData(SCT_obj, 
                       metadata =  Idents(SCT_obj), 
                       col.name = "clusters_NPC")
SCT_obj$clusters_NPC <- factor(SCT_obj$clusters_NPC, levels = clusters_order[2:15])

UMAPPlot(SCT_obj, group.by = "clusters_NPC", label = T) + NoLegend()
```


```{r saving object}
saveRDS(object = SCT_obj, file = "Objects/SCT_obj.rds")
sessionInfo()
```
