==> 6_Dcx_analysis.Rmd <==
---
title: "Dcx classes creation and irradiation ratios"
output: html_document
---

Here we define DcxHigh and DcxLow cells amongst s-iNB and examine their relative abundance
```{r Libraries}
library(tidyverse)
library(dplyr)
library(patchwork)
library(Seurat, lib.loc = "/opt/rstudio-server_conda/conda/envs/rstudio-server_4.1.0/lib/R/library/")
```

Loading required objects
This analysis is focused on NPC populations, therefore we will use a subset of the base object
```{r Import objects}
SCT_obj <- readRDS("Objects/SCT_obj.rds")
Biomart_annotated <- read.csv("Analysis_tools/Biomart_annotated.csv")
geneIDs <- # For easy calling of genes through external gene name
  Biomart_annotated[!duplicated(Biomart_annotated$external_gene_name),1]
names(geneIDs) <- Biomart_annotated[!duplicated(Biomart_annotated$external_gene_name),2]
NPC_obj <- subset(SCT_obj, NPC == "NPC")
```

```{r Fetch Dcx expression values}
dcx_expression <- FetchData( #FetchData sur le slot data
    NPC_obj,
    vars = c(geneIDs["Dcx"], "ident", "Status"),
    assay = "SCT",
    slot = "counts"
  ) %>%
    mutate(dcx = ifelse(ENSMUSG00000031285 > 1,"DcxHigh","DcxLow"))
dcx_lvl <- dcx_expression$dcx
names(dcx_lvl) <- rownames(dcx_expression)
SCT_obj <- AddMetaData(SCT_obj, metadata = dcx_lvl, col.name = "dcx_lvl")

colnames(dcx_expression)[1:4] <- c("DCX_val", "cluster", "Status", "dcx")  

write.csv(table(dcx_expression$cluster, 
                dcx_expression$Status,
                dcx_expression$dcx),
          file = "Outputs/DCX_cluster_status.csv")
```

Saving object with all metadata
```{r}
saveRDS(object = SCT_obj, file = "Objects/Ber23_SVZ_IRR.rds")
sessionInfo()
```
